My lighter version of dot-rot. TODO: Script to clone .emacs.d in.

+++ lib/dotrot.rb
@@ -14,24 +14,28 @@ class KeyBindingsInstaller
   end
 end
 
-def symlink_dotfile(name)
+def symlink_dotfile(name, dry_run = false)
   source = File.join(DOTROT_HOME, "dot.#{name}")
   target = File.join(ENV['HOME'], ".#{name}")
   
-  backup_dotfile(target)
+  puts "** backing up #{name}..."
+  backup_dotfile(target, dry_run)
 
   puts "** symlinking #{name}... "
-  FileUtils.ln_s source, target
+  FileUtils.ln_s source, target unless dry_run
   puts '** done!'
   puts
 end
 
-def backup_dotfile(target)
+def backup_dotfile(target, dry_run = false)
   if File.exists?(target)
     name = File.basename(target)
     puts "** moving old #{name} to #{DOTROT_BACKUPS}/#{name}"
-    FileUtils.rm_rf(File.join(DOTROT_BACKUPS, name))
-    FileUtils.mkdir_p(DOTROT_BACKUPS)
-    FileUtils.mv target, File.join(DOTROT_BACKUPS, name), :force => true
+
+    unless dry_run
+      FileUtils.rm_rf(File.join(DOTROT_BACKUPS, name))
+      FileUtils.mkdir_p(DOTROT_BACKUPS)
+      FileUtils.mv target, File.join(DOTROT_BACKUPS, name), :force => true
+    end
   end
 end
